+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change log
 
 ## 0.3.0 
-[2019-Mar-9]
+[2019-Mar-10]
 - Add [react-dropzone](https://react-dropzone.js.org/). It is able to upload test report.
 - Remove static test file.
+- Add test case visibility filter.
 
 ## 0.2.0  
 [2018-Nov-25]